requirements.txt
absl-py==2.1.0
antlr4-python3-runtime==4.9.3
asttokens==2.4.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
decorator==5.1.1
exceptiongroup==1.2.2
executing==2.1.0
filelock==3.13.1
Fiona==1.9.4.post1
fonttools==4.55.0
fsspec==2024.2.0
geopandas==0.13.0
gitdb==4.0.11
GitPython==3.1.43
grpcio==1.68.0
hydra-core==1.3.2
idna==3.10
ipython==8.29.0
jedi==0.19.2
Jinja2==3.1.3
kiwisolver==1.4.7
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
omegaconf==2.3.0
opencv-python==4.10.0.84
osmnx==1.3.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==10.2.0
prompt_toolkit==3.0.48
protobuf==5.28.3
psutil==6.1.0
pure_eval==0.2.3
Pygments==2.18.0
pyparsing==3.2.0
pyproj==3.5.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
scipy==1.14.1
seaborn==0.13.2
shapely==2.0.1
six==1.16.0
smmap==5.0.1
stack-data==0.6.3
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
thop==0.1.1.post2209072238
torch==2.5.1+cpu
torchaudio==2.5.1+cpu
torchvision==0.20.1+cpu
tqdm==4.67.0
traitlets==5.14.3
typing_extensions==4.9.0
tzdata==2024.2
ultralytics==8.3.36
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.1.3